% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-pdo.R
\name{download_pdo}
\alias{download_pdo}
\title{Download PDO Index Data}
\usage{
download_pdo()
}
\value{
A tbl data.frame of the PDO index data.
}
\description{
Downloads PDO index data from
\url{http://research.jisao.washington.edu/pdo/PDO.latest}.
}
\details{
For more information see \url{https://github.com/poissonconsulting/rpdo}.
}

