% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdo-download.R
\name{pdo_download}
\alias{pdo_download}
\alias{download_pdo}
\title{Download PDO Index Data}
\usage{
pdo_download()

download_pdo()
}
\value{
A data frame of the PDO index data.
}
\description{
Downloads PDO index data from the no longer updated
\url{http://research.jisao.washington.edu/pdo/PDO.latest}.
}
\details{
\lifecycle{soft-deprecated}

Deprecated for \code{\link[rsoi:download_pdo]{rsoi::download_pdo()}}.

\code{download_pdo()} is an alias for \code{pdo_download()}.
}
