% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rpdo-package}
\alias{rpdo-package}
\title{Pacific Decadal Oscillation Index Data}
\description{
Monthly Pacific Decadal Oscillation (PDO) index
values from January 1900.
}
\details{
Includes \code{\link{pdo_download}()} to scrape the latest values from
\url{http://research.jisao.washington.edu/pdo/PDO.latest}.
}
