% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rpdo}
\alias{rpdo}
\alias{rpdo-package}
\title{Pacific Decadal Oscillation Index Data}
\description{
Monthly Pacific Decadal Oscillation (PDO) index
values from January 1900 to present.
}
\details{
For more information see \url{https://github.com/poissonconsulting/rpdo}.
}

