% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdo-download.R
\name{pdo_download}
\alias{pdo_download}
\alias{download_pdo}
\title{Download PDO Index Data}
\usage{
pdo_download()

download_pdo()
}
\value{
A data frame of the PDO index data.
}
\description{
Downloads PDO index data from
\url{http://research.jisao.washington.edu/pdo/PDO.latest}.
}
\details{
\code{download_pdo()} is an alias for \code{pdo_download()}.

For more information see \url{https://github.com/poissonconsulting/rpdo}.
}
