% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpdo-package.r
\docType{package}
\name{rpdo-package}
\alias{rpdo}
\alias{rpdo-package}
\title{rpdo: Pacific Decadal Oscillation Index Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Monthly Pacific Decadal Oscillation (PDO) index values from January 1900 to September 2018. Superseded by 'rsoi' package which includes the historical and most recent monthly PDO index values together with related climate indices.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/poissonconsulting/rpdo}
  \item Report bugs at \url{https://github.com/poissonconsulting/rpdo/issues}
}

}
\author{
\strong{Maintainer}: Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})

Authors:
\itemize{
  \item Nathan Mantua \email{nate.mantua@noaa.gov} [data contributor]
  \item Steven R. Hare \email{hare@iphc.washington.edu} [data contributor]
}

}
\keyword{internal}
