% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpdo-package.r
\docType{package}
\name{rpdo}
\alias{rpdo}
\title{rpdo}
\description{
Monthly Pacific Decadal Oscillation (PDO) index
values from January 1900.
}
\seealso{
\code{\link[=pdo]{pdo()}} and \code{\link[=pdo_download]{pdo_download()}}
}
